Feito inserir (Cliente e Pedido) e listar (Pizza, Cliente)

+++ Pizzaria_Marchiori/src/main/java/Dominio/Cliente.java
@@ -100,7 +100,7 @@ public class Cliente implements Serializable {
         this.rua = rua;
     }
 
-    public int getNumero() {
+    public float getNumero() {
         return numero;
     }
